Initial Commit : imagine - an image steganographic tool
The initial commit for our new product imagine. It's a steganographic
tool for hiding data in the pixel data of images. It forgoes the file
system capabilities of stegfs, instead opting for more covert use
alongside encrypt.

The expected use case goes something like this: encrypt your data
(usinng encrypt, obviously) and then embed it in an image of your
choice.

It currently only supports PNG and TIFF images; we're considering JPEG,
but their lossy nature is putting us off. (I personally don't use JPEG
for my photos, preferring either PNG or TIFF depending on whether I'm
processing them or passing them around.)

The data is hidden in the least-significant bits of the RGB pixel
values: 3 bits in both the red and blue channels and only 2 in the
green; therefore imagine is able to store 1 byte per pixel. We
experimented with more data per pixel but it become obvious the image
had been tampered with.