UI Update
Add command line progress bars (from common). First bar is current task,
second is overall progress. Modules use a callback to update values as
this should hopefully make updating the GUI easier. And on that note,
the GUI has been simplified, and there's been some work to plumb it in.