Add a truly steganographic option!
If you specify -f as the first argument to hide, it will not store the
length of the data hidden (you will need to specify it again to get the
data back); it will then hide the file and fill the remaining bits with
random values. If you use encrypt and the --raw option you data can be
truly hidden in addition to being encrypted. encrypt is smart enough to
ignore extra random data beyond the end of the encrypted archive (it
encrypted the length of encrypted data).