Add helper script
This script will encrypt a file before hiding it, and decrypt it again
when recovering it. It uses the --raw option for encrypt, which causes
encrypt not to generate a header, and the -f fill option for hide, which
does not prefix the file size to the hidden data as well as pad with
random noise the entire image.